backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==14.1
pydantic==2.9.2
pydantic-settings==2.6.0

# Database (shared with main VoiceKit)
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Google AI (for web_search and deep_research skills)
google-genai>=1.0.0
# Google ID token verification (used by auth.py for JWT validation)
google-auth>=2.0.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0

# CORS support
python-multipart==0.0.12

# MCP (Model Context Protocol) client
mcp>=1.0.0
